Move my-identity into commons
